create fun.py file and place related code to this page from main.py and refactor code
